+++ src/HFRouter.js
@@ -1,11 +1,11 @@
 import React from 'react'
 import { Route } from 'react-router-dom'
 
-import Dashboard from 'pages/holofuel/Dashboard'
-import TransactionHistory from 'pages/holofuel/TransactionHistory'
-import Inbox from 'pages/holofuel/Inbox'
-import CreateOffer from 'pages/holofuel/CreateOffer'
-import CreateRequest from 'pages/holofuel/CreateRequest'
+import Dashboard from 'holofuel/pages/Dashboard'
+import TransactionHistory from 'holofuel/pages/TransactionHistory'
+import Inbox from 'holofuel/pages/Inbox'
+import CreateOffer from 'holofuel/pages/CreateOffer'
+import CreateRequest from 'holofuel/pages/CreateRequest'
 
 export default function HFRouter () {
   return <>
