+++ src/holofuel/components/HashIcon/HashIcon.js
