+++ src/holofuel/components/HashIcon/HashIcon.module.css
