+++ src/holofuel/components/HashIcon/index.js
