+++ src/holofuel/pages/CreateOffer/CreateOffer.test.js
@@ -7,7 +7,7 @@ import CreateOffer, { FEE_PERCENTAGE } from './CreateOffer'
 import { TYPE } from 'models/Transaction'
 import HolofuelOfferMutation from 'graphql/HolofuelOfferMutation.gql'
 
-jest.mock('components/holofuel/Header')
+jest.mock('holofuel/components/Header')
 
 const counterparty = 'HcScic3VAmEP9ucmrw4MMFKVARIvvdn43k6xi3d75PwnOswdaIE3BKFEUr3eozi'
 const amount = 35674
