+++ src/holofuel/pages/TransactionHistory/TransactionHistory.js
@@ -5,9 +5,9 @@ import _ from 'lodash'
 import { useQuery, useMutation } from '@apollo/react-hooks'
 import { isEmpty } from 'lodash/fp'
 import './TransactionHistory.module.css'
-import PrimaryLayout from 'components/holofuel/layout/PrimaryLayout'
-import Button from 'components/holofuel/Button'
-import Modal from 'components/holofuel/Modal'
+import PrimaryLayout from 'holofuel/components/layout/PrimaryLayout'
+import Button from 'holofuel/components/Button'
+import Modal from 'holofuel/components/Modal'
 
 import HolofuelCompletedTransactionsQuery from 'graphql/HolofuelCompletedTransactionsQuery.gql'
 import HolofuelWaitingTransactionsQuery from 'graphql/HolofuelWaitingTransactionsQuery.gql'
